+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## 0.1.0
+## 0.2.0 - 2022-05-28
+
+### Added
+- Also cancel periodicTimers, not just simple timers.
+- New `CancellableContext` class and `CancellableFuture.ctx` constructor.
+- New `ctx.scheduleOnCancel` function to run callbacks on cancellation.
+- Added `currentCancellableContext()` function.
+
+## 0.1.0 - 2022-05-27
 
 - Initial version.